app-emulation/qemu: Create new udev file for VFIO
Create a new udev file for VFIO devices.  With current settings,
permissions are set up as:

    drwxr-xr-x 2 root root 80 Mar 25 18:38 /dev/vfio
    crw------- 1 root root 247,   0 Mar 25 18:38 /dev/vfio/1
    crw-rw-rw- 1 root root  10, 196 Mar 25 18:37 /dev/vfio/vfio

This is potentially unsafe as it allows any user to write to
/dev/vfio/vfio.  It also restricts access to /dev/vfio/1 to root which
prevents non-root users from performing vga-passthrough.  The new rules
file, however, alters the permissions to become the following:

    drwxr-xr-x 2 root root 80 Mar 25 18:38 /dev/vfio
    crw-rw---- 1 root kvm 247,   0 Mar 25 19:04 /dev/vfio/1
    crw-rw---- 1 root kvm  10, 196 Mar 25 19:04 /dev/vfio/vfio

This relaxes permissions on /dev/vfio/1 to allow any user of the kvm
group to perform vga-passthrough.  Furthermore, it tighens the
permissions on /dev/vfio/vfio to restrict access to only kvm users.